Extracted unique id generation into its own module
